+++ contracts/test/TestCallbackWithRevert.sol
@@ -25,7 +25,7 @@ contract TestCallbackWithRevert {
         soRandom = _soRandom;
     }
 
-    function soRandomCallback(uint256 _id, bytes32 value) external {
+    function soRandomCallback(uint128 _id, bytes32 value) external {
         revert("Revert");
     }
